Improved map commands a bit.
