+++ SS14.Server/SS14.Server.csproj
@@ -113,6 +113,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Reflection\ServerReflectionManager.cs" />
     <Compile Include="BaseServer.cs" />
+    <Compile Include="ServerConsole\Commands\MapCommands.cs" />
     <Compile Include="Signals.cs" />
     <Content Include="server_config.toml">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
